Fixed a few 'signed/unsigned mismatch' compiler warnings in   cssys/csuctransform.h. Re-enabled 'signed/unsigned mismatch' compiler warnings for MSVC.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15654 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/cssys/csuctransform.h
@@ -53,7 +53,7 @@ public:
     return chUsed;
   
 #define GET_NEXT(next)	\
-  if (chUsed == strlen)				\
+  if ((size_t)chUsed == strlen)			\
   {						\
     FAIL(chUsed);				\
   }						\
@@ -731,7 +731,7 @@ public:
     
     // Skip backward to the first byte of the sequence.
     int skip = 1;
-    while (((*pos & 0xc0) == 0x80) && (skip < maxRew))
+    while (((*pos & 0xc0) == 0x80) && ((size_t)skip < maxRew))
     {
       skip++;
       pos--;
